+++ jwp-basic-gradle/src/main/java/core/mvcframework/DispatcherServlet.java
@@ -1,5 +1,6 @@
 package core.mvcframework;
 
+import core.mvcframework.controller.Controller;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
