+++ jwp-basic-gradle/src/main/java/core/mvcframework/RequestMapping.java
@@ -1,5 +1,7 @@
 package core.mvcframework;
 
+import core.mvcframework.controller.Controller;
+import core.mvcframework.controller.ForwardController;
 import java.util.HashMap;
 import java.util.Map;
 import next.controller.HomeController;
