+++ jwp-basic-gradle/src/main/java/core/mvcframework/controller/Controller.java
@@ -1,4 +1,4 @@
-package core.mvcframework;
+package core.mvcframework.controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
