+++ jwp-basic-gradle/src/main/java/core/mvcframework/controller/ForwardController.java
@@ -1,5 +1,6 @@
-package core.mvcframework;
+package core.mvcframework.controller;
 
+import core.mvcframework.controller.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
