+++ jwp-basic-gradle/src/main/java/next/controller/qna/ShowController.java
@@ -1,6 +1,6 @@
 package next.controller.qna;
 
-import core.mvcframework.Controller;
+import core.mvcframework.controller.Controller;
 import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
