+++ jwp-basic-gradle/src/main/java/next/controller/user/ListUserController.java
@@ -1,6 +1,6 @@
 package next.controller.user;
 
-import core.mvcframework.Controller;
+import core.mvcframework.controller.Controller;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
